+++ config/software/graphicsmagick.rb
@@ -23,6 +23,7 @@ license_file 'Copyright.txt'
 skip_transitive_dependency_licensing true
 
 dependency 'libpng'
+dependency 'libjpeg-turbo'
 dependency 'zlib'
 
 source url: "https://ftp.icm.edu.pl/pub/unix/graphics/GraphicsMagick/1.3/GraphicsMagick-#{version}.tar.gz",
@@ -45,20 +46,20 @@ build do
     '--without-gslib',
     '--without-jbig',
     '--without-webp',
-    '--without-jpeg',
     '--without-jp2',
     '--without-lcms2',
     '--without-lzma',
-    '--with-png',
-    "--with-sysroot=#{install_dir}/embedded",
     '--without-tiff',
     '--without-trio',
     '--without-ttf',
     '--without-umem',
     '--without-wmf',
     '--without-xml',
-    '--with-zlib',
-    '--without-x'
+    '--without-x',
+    '--with-jpeg=yes',
+    '--with-zlib=yes',
+    '--with-png=yes',
+    "--with-sysroot=#{install_dir}/embedded"
   ]
 
   command configure_command.join(' '), env: env