+++ files/gitlab-cookbooks/gitlab/libraries/gitlab_rails.rb
@@ -43,6 +43,7 @@ module GitlabRails # rubocop:disable Style/MultilineIfModifier
       parse_uploads_dir
       parse_packages_dir
       parse_dependency_proxy_dir
+      parse_terraform_state_dir
       parse_pages_dir
       parse_repository_storage
     end
@@ -177,6 +178,11 @@ module GitlabRails # rubocop:disable Style/MultilineIfModifier
     def parse_dependency_proxy_dir
       # This requires the parse_shared_dir to be executed before
       Gitlab['gitlab_rails']['dependency_proxy_storage_path'] ||= File.join(Gitlab['gitlab_rails']['shared_path'], 'dependency_proxy')
+    end
+
+    def parse_terraform_state_dir
+      # This requires the parse_shared_dir to be executed before
+      Gitlab['gitlab_rails']['terraform_state_storage_path'] ||= File.join(Gitlab['gitlab_rails']['shared_path'], 'terraform_state')
     end
 
     def parse_pages_dir