+++ spec/chef/recipes/gitlab-rails_spec.rb
@@ -70,6 +70,10 @@ describe 'gitlab::gitlab-rails' do
       expect(chef_run).not_to run_ruby_block('directory resource: /tmp/shared/dependency_proxy')
     end
 
+    it 'does not create the terraform_state storage directory' do
+      expect(chef_run).not_to run_ruby_block('directory resource: /tmp/shared/terraform_state')
+    end
+
     it 'does not create the uploads storage directory' do
       expect(chef_run).not_to run_ruby_block('directory resource: /tmp/uploads')
     end
@@ -117,6 +121,10 @@ describe 'gitlab::gitlab-rails' do
 
     it 'creates the dependency_proxy directory' do
       expect(chef_run).to create_storage_directory('/tmp/shared/dependency_proxy').with(owner: 'git', mode: '0700')
+    end
+
+    it 'creates the terraform_state directory' do
+      expect(chef_run).to create_storage_directory('/tmp/shared/terraform_state').with(owner: 'git', mode: '0700')
     end
 
     it 'creates the uploads directory' do
@@ -615,6 +623,44 @@ describe 'gitlab::gitlab-rails' do
       end
     end
 
+    context 'for settings regarding object storage for terraform_state' do
+      it 'allows not setting any values' do
+        expect(chef_run).to create_templatesymlink('Create a gitlab.yml and create a symlink to Rails root').with_variables(
+          hash_including(
+            'terraform_state_storage_path' => '/var/opt/gitlab/gitlab-rails/shared/terraform_state',
+            'terraform_state_object_store_enabled' => false,
+            'terraform_state_object_store_remote_directory' => 'terraform_state'
+          )
+        )
+      end
+
+      context 'with values' do
+        before do
+          stub_gitlab_rb(gitlab_rails: {
+                           terraform_state_object_store_enabled: true,
+                           terraform_state_object_store_direct_upload: true,
+                           terraform_state_object_store_background_upload: false,
+                           terraform_state_object_store_proxy_download: true,
+                           terraform_state_object_store_remote_directory: 'mepmep',
+                           terraform_state_object_store_connection: aws_connection_hash
+                         })
+        end
+
+        it "sets the object storage values" do
+          expect(chef_run).to create_templatesymlink('Create a gitlab.yml and create a symlink to Rails root').with_variables(
+            hash_including(
+              'terraform_state_object_store_enabled' => true,
+              'terraform_state_object_store_direct_upload' => true,
+              'terraform_state_object_store_background_upload' => false,
+              'terraform_state_object_store_proxy_download' => true,
+              'terraform_state_object_store_remote_directory' => 'mepmep',
+              'terraform_state_object_store_connection' => aws_connection_hash
+            )
+          )
+        end
+      end
+    end
+
     describe 'pseudonymizer settings' do
       it 'allows not setting any values' do
         expect(chef_run).to create_templatesymlink('Create a gitlab.yml and create a symlink to Rails root').with_variables(
@@ -960,6 +1006,34 @@ describe 'gitlab::gitlab-rails' do
             )
           )
         end
+      end
+    end
+
+    context 'when seat link is enabled' do
+      it 'sets seat link to true' do
+        expect(chef_run).to create_templatesymlink('Create a gitlab.yml and create a symlink to Rails root').with_variables(
+          hash_including(
+            'seat_link_enabled' => true
+          )
+        )
+      end
+    end
+
+    context 'when seat link is disabled' do
+      before do
+        stub_gitlab_rb(
+          gitlab_rails: {
+            seat_link_enabled: false,
+          }
+        )
+      end
+
+      it 'sets seat link to false' do
+        expect(chef_run).to create_templatesymlink('Create a gitlab.yml and create a symlink to Rails root').with_variables(
+          hash_including(
+            'seat_link_enabled' => false
+          )
+        )
       end
     end