+++ CHANGELOG.md
@@ -3,6 +3,28 @@
 The latest version of this file can be found at the master branch of the
 omnibus-gitlab repository.
 
+## 12.9.2 (2020-03-31)
+
+### Fixed (1 change)
+
+- Configures logrotate service for puma. !4024
+
+### Added (1 change)
+
+- Allow setting in seat_link_enabled in gitlab.rb. !4042
+
+### Other (1 change)
+
+- Update Mattermost to 5.20.2.
+
+
+## 12.9.1 (2020-03-26)
+
+### Security (1 change)
+
+- Bump pcre2 version to 10.34.
+
+
 ## 12.9.0 (2020-03-22)
 
 ### Fixed (5 changes, 1 of them is from the community)
@@ -115,6 +137,13 @@ omnibus-gitlab repository.
 - Bump PostgreSQL versions to 9.6.17, 10.12, and 11.7. !3933
 - Upgrade Mattermost to 5.19.1.
 - Update Mattermost to 5.18.1.
+
+
+## 12.7.8 (2020-03-26)
+
+### Security (1 change)
+
+- Bump pcre2 version to 10.34.
 
 
 ## 12.7.7